Css inicial incluyendo la tabla de fondos

+++ src/Components/TableRow.js
@@ -1,31 +1,34 @@
 import React,{useEffect} from 'react'
+import styles from "../CSS/TableRow.module.css"
 
 export default function TableRow(props) {
 
 
     const handleDeleteData = () =>{
-        props.deleteData(props.id);
 
-        // sumar el valor del fondo que se eliminar
-        let newData = props.fondos.filter((el)=>el.id !== props.id)
-        let valor = 0;
-        newData.forEach(e=>valor+=e.monto)
-        props.setCapitalInicial((el)=>500000-valor)
+        let borrarData = window.confirm("¿eliminar este registro");
+        if(borrarData){
+
+            props.deleteData(props.id);
+            
+            // sumar el valor del fondo que se eliminar
+            let newData = props.fondos.filter((el)=>el.id !== props.id)
+            let valor = 0;
+            newData.forEach(e=>valor+=e.monto)
+            props.setCapitalInicial((el)=>500000-valor)
+        }
+        else{
+            return;
+        }
     } 
-    
-    // useEffect(()=>{
-        
-    //     console.log(props.fondos)
-        
-    // },[props.fondos])
     
 
     return(
-        <tr>
+        <tr className={styles.table_row}>
             <td>{props.nombre}</td>
             <td>{props.monto}</td>
             <td>{props.categoria}</td>
-            <td><button onClick={handleDeleteData}>Desvincular Fondo</button></td>
+            <td><button className={styles.btn} onClick={handleDeleteData}>Desvincular Fondo</button></td>
         </tr>
     );
